microservices/simple/onedrive/requirements.txt
requests
flask
azure-identity
msgraph-sdk